+++ todo-backend/package.json
@@ -13,15 +13,16 @@
   "license": "ISC",
   "description": "",
   "dependencies": {
+    "@azure/cosmos": "^4.3.0",
     "@azure/monitor-opentelemetry": "^1.9.0",
+    "@azure/opentelemetry-instrumentation-azure-sdk": "^1.0.0-beta.8",
     "@opentelemetry/instrumentation": "^0.200.0",
     "@opentelemetry/instrumentation-express": "^0.48.0",
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
     "express": "^5.1.0",
     "swagger-jsdoc": "^6.2.8",
-    "swagger-ui-express": "^5.0.0",
-    "@azure/cosmos": "^4.3.0"
+    "swagger-ui-express": "^5.0.0"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.26.9",
